Feat: 팔로우 버튼을 사용하는 곳에서 follow상태를 response하도록 수정

+++ src/main/java/com/finns/follow/mapper/FollowMapper.java
@@ -41,7 +41,7 @@ public interface FollowMapper {
      * @param user_no 팔로우 정보를 담은 DTO
      * @return 팔로우 관계가 존재하면 1, 존재하지 않으면 0
      */
-    int checkFollowExists(@Param("user_no") int user_no, @Param("to_user_no") int to_user_no);
+    int checkFollowExists(@Param("user_no") long user_no, @Param("to_user_no") long to_user_no);
 
     int countFollowers(int user_no);
     int countFollowing(int user_no);
